Split page editor to component
