+++ framerail/src/routes/[slug]/[...extra]/index.ts
@@ -1,3 +1,4 @@
+export { default as EditorPane } from "./EditorPane.svelte"
 export { default as LayoutPane } from "./LayoutPane.svelte"
 export { default as MovePane } from "./MovePane.svelte"
 export { default as ParentPane } from "./ParentPane.svelte"
