+++ framerail/src/routes/[slug]/[...extra]/page.svelte
@@ -4,7 +4,7 @@
   import { onMount } from "svelte"
   import { useErrorPopup, usePagePaneState } from "$lib/stores"
   import { PagePane } from "$lib/types"
-  import { LayoutPane, MovePane, ParentPane } from "."
+  import { EditorPane, LayoutPane, MovePane, ParentPane } from "."
   let showErrorPopup = useErrorPopup()
   let pagePaneState = usePagePaneState()
 
@@ -55,38 +55,6 @@
     goto(`/${$page.data.page.slug}${options.join("")}/edit`, {
       noScroll: true
     })
-  }
-
-  function cancelEdit() {
-    let options: string[] = []
-    if ($page.data.options.no_render) options.push("norender")
-    options = options.map((opt) => `/${opt}`)
-    goto(`/${$page.data.page.slug}${options.join("")}`, {
-      noScroll: true
-    })
-  }
-
-  async function saveEdit() {
-    let form = document.getElementById("editor")
-    let fdata = new FormData(form)
-    fdata.set("site-id", $page.data.site.site_id)
-    fdata.set("page-id", $page.data.page.page_id)
-    fdata.set("last-revision-id", $page.data.page_revision.revision_id)
-    let res = await fetch(`/${$page.data.page.slug}/edit`, {
-      method: "POST",
-      body: fdata
-    }).then((res) => res.json())
-    if (res?.message) {
-      showErrorPopup.set({
-        state: true,
-        message: res.message,
-        data: res.data
-      })
-    } else {
-      goto(`/${$page.data.page.slug}`, {
-        noScroll: true
-      })
-    }
   }
 
   async function handleHistory() {
@@ -503,51 +471,7 @@
 </div>
 
 {#if $page.data.options?.edit}
-  <form id="editor" class="editor" method="POST" on:submit|preventDefault={saveEdit}>
-    <input
-      name="title"
-      class="editor-title"
-      placeholder={$page.data.internationalization?.title}
-      type="text"
-      value={$page.data.page_revision.title}
-    />
-    <input
-      name="alt-title"
-      class="editor-alt-title"
-      placeholder={$page.data.internationalization?.["alt-title"]}
-      type="text"
-      value={$page.data.page_revision.alt_title}
-    />
-    <textarea name="wikitext" class="editor-wikitext">{$page.data.wikitext}</textarea>
-    <input
-      name="tags"
-      class="editor-tags"
-      placeholder={$page.data.internationalization?.tags}
-      type="text"
-      value={$page.data.page_revision.tags.join(" ")}
-    />
-    <textarea
-      name="comments"
-      class="editor-comments"
-      placeholder={$page.data.internationalization?.["wiki-page-revision-comments"]}
-    />
-    <div class="action-row editor-actions">
-      <button
-        class="action-button editor-button button-cancel clickable"
-        type="button"
-        on:click|stopPropagation={cancelEdit}
-      >
-        {$page.data.internationalization?.cancel}
-      </button>
-      <button
-        class="action-button editor-button button-save clickable"
-        type="submit"
-        on:click|stopPropagation
-      >
-        {$page.data.internationalization?.save}
-      </button>
-    </div>
-  </form>
+  <EditorPane />
 {:else}
   <div class="action-row editor-actions">
     <button
@@ -1244,7 +1168,6 @@
     text-align: right;
   }
 
-  .editor,
   .file-upload,
   .file-edit,
   .file-move,
@@ -1255,10 +1178,6 @@
     align-items: stretch;
     justify-content: stretch;
     width: 80vw;
-  }
-
-  .editor-wikitext {
-    height: 60vh;
   }
 
   .page-source,
